Default to gemini-2.0-flask
